Refactored IIN Agent Code
The IIN Agent is now a single server offering multiple services and activating one or more DLT extensions.
