+++ core/identity-management/iin-agent/.env.template
@@ -3,3 +3,4 @@ IIN_AGENT_TLS=false
 IIN_AGENT_TLS_CERT_PATH=path_to_tls_cert_pem_for_driver
 IIN_AGENT_TLS_KEY_PATH=path_to_tls_key_pem_for_driver
 NETWORK_NAME=network1
+DLT_TYPE=<fabric|corda|besu|...>
