+++ core/identity-management/iin-agent/.npmrc.template
