--- core/identity-management/iin-agent/common/README.md
@@ -1,8 +1,0 @@
-<!--
- Copyright IBM Corp. All Rights Reserved.
-
- SPDX-License-Identifier: CC-BY-4.0
- -->
-# IIN Agent Implementation
-
-In this folder lies an implementation of the core, or common, part of an IIN agent according to the [RFC specification](https://github.com/hyperledger-labs/weaver-dlt-interoperability/blob/main/rfcs/models/identity/iin-agent.md). This module is built in a DLT-neutral manner and can be used by a participant of a network running on any DLT platform.
