+++ core/identity-management/iin-agent/src/protocols/README.md
@@ -1,0 +1,22 @@
+<!--
+ Copyright IBM Corp. All Rights Reserved.
+
+ SPDX-License-Identifier: CC-BY-4.0
+ -->
+# Agent Protocols and Services
+
+## Syncing External Network State from Foreign IIN Agents: `syncExternalState`
+
+_TODO_
+
+## Flow Through Local IIN Agents for Attestation Consensus: `flowAndRecordAttestations`
+
+_TODO_
+
+## Serving Attestation Requests from Local IIN Agents: `requestAttestation`
+
+_TODO_
+
+## Processing Attestations from Local IIN Agents: `sendAttestation`
+
+_TODO_
