+++ core/identity-management/iin-agent/tsconfig.json
